% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnadc_download.R
\name{pnadc_download}
\alias{pnadc_download}
\title{Download PNADc}
\usage{
pnadc_download(year, quartile, path = NULL)
}
\arguments{
\item{year}{The year of the PNADc edition to download. Must be a number between 2012 and the current year.}

\item{quartile}{The quartile of the year to download. Must be a number between 1 and 4.}

\item{path}{Path of the local directory where the PNADc files are stored or where you want to save the data.}
}
\description{
Downloads the PNADc data and creates a survey design on your computer.
It saves an external representation of an R object that can be later read in other functions of PNADc.table.
}
\examples{
\dontrun{
pnadc_download(2018, 1)
}
}
